workspaces/sasoli0502/todoist-sync/pyproject.toml
[project]
name = "todoist-sync"
version = "0.1.0"
description = "Sync CLAUDE.md tasks to Todoist"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
]
